Fix handling of images smaller than lowest breakpoint
When selecting an image within the response media field that's smaller
than the smallest breakpoint, the image is ignored and nothing is
displayed. To improve this and always show an image, the content editor
is displayed the image and a message stating that it's smaller than the
smallest breakpoint and this image will be displayed in an `img` tag
on the front-end.

Fixes #46